==> contrastive/notebooks/ipmi_2023/03-analysis_nearest-neighbours-extended.ipynb <==
Notebook to compute the nn distance between latent spaces.

import os
import torch
import json
import numpy as np
import numpy.random as rd
import pandas as pd
import matplotlib.pyplot as plt
plt.style.use('default')

from pandas.plotting import scatter_matrix
from scipy.spatial import distance
from scipy.stats import percentileofscore

from contrastive.utils.nn_distance import *

## Load the embeddings

path1 = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/pca/30/Run1/pca_embeddings.csv"

embeddings1 = pd.read_csv(path1, index_col=0)

print(embeddings1.head())

embeddings2 = -embeddings1
embeddings2.head()

embeddings1.loc['1-17COLMA_t0', :]

# get the subject list

# small check before
if embeddings1.index.all() != embeddings2.index.all():
    raise ValueError("The two spaces don't contain the same subjects")
else:
    list_subjects = embeddings1.index.astype('str')

list_subjects

## Compute the space distances

dist_mat_1 = get_distance_matrix(embeddings1)
dist_mat_2 = get_distance_matrix(embeddings2)

dist_mat_1['1-03HUBJO_t0'].sort_values().index

perc_mat_1 = get_percentile_matrix(dist_mat_1)
perc_mat_2 = get_percentile_matrix(dist_mat_2)
perc_mat_1

get_distance(perc_mat_1, perc_mat_2)

latent_space_distance(embeddings1, embeddings2)

## Apply the method to collected data

path1 = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/pca/30/Run1/pca_embeddings.csv"
path2 = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/pca/30/Run2/pca_embeddings.csv"

embeddings1 = pd.read_csv(path1, index_col=0)
embeddings2 = pd.read_csv(path2, index_col=0)

latent_space_distance(embeddings1, embeddings2)

# for a good net
path1 = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/convnet/no_foldlabel_30/13-15-48/cingulate_ACCpatterns_embeddings/full_embeddings.csv"
path2 = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/convnet/no_foldlabel_30/16-39-35/cingulate_ACCpatterns_embeddings/full_embeddings.csv"

embeddings1 = pd.read_csv(path1, index_col=0)
embeddings2 = pd.read_csv(path2, index_col=0)

latent_space_distance(embeddings1, embeddings2)

# for a bad net
path1 = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/densenet2/no_foldlabel_4/11-41-24/cingulate_ACCpatterns_embeddings/full_embeddings.csv"
path2 = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/densenet2/no_foldlabel_4/15-20-30/cingulate_ACCpatterns_embeddings/full_embeddings.csv"

embeddings1 = pd.read_csv(path1, index_col=0)
embeddings2 = pd.read_csv(path2, index_col=0)

latent_space_distance(embeddings1, embeddings2)

# compute the space distances for some model types and functions

paths1 = ["/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/pca/30/Run1/pca_embeddings.csv",
          "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/convnet/no_foldlabel_30/13-15-48/cingulate_ACCpatterns_embeddings/full_embeddings.csv",
          "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/densenet2/no_foldlabel_4/11-41-24/cingulate_ACCpatterns_embeddings/full_embeddings.csv"]
paths2 = ["/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/pca/30/Run2/pca_embeddings.csv",
          "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/convnet/no_foldlabel_30/16-39-35/cingulate_ACCpatterns_embeddings/full_embeddings.csv",
          "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/densenet2/no_foldlabel_4/15-20-30/cingulate_ACCpatterns_embeddings/full_embeddings.csv"]

idx_names = ['same', 'pca', 'good_net', 'bad_net', 'random']

functions = [np.sqrt, custom_ln, custom_identity]
functions_names = ['sqrt', 'custom_ln', 'custom_identity']

n_rows = len(idx_names)
n_fct = len(functions) + 1

results_df = pd.DataFrame(np.zeros((n_rows, n_fct)), columns=['original_nn']+functions_names, 
                          index=idx_names)


for j,fct in enumerate(functions):
    print(functions_names[j])
    for i in range(len(paths1)):
        embs1 = pd.read_csv(paths1[i], index_col=0)
        embs2 = pd.read_csv(paths2[i], index_col=0)
        value = latent_space_distance(embs1, embs2, fct=fct)
        print(i+1,value)
        results_df[functions_names[j]][idx_names[i+1]] = value

# random embeddings
random_embs_1 = pd.DataFrame(np.random.normal(size=(341,30)))
random_embs_2 = pd.DataFrame(np.random.normal(size=(341,30)))

for j,fct in enumerate(functions):
    print(functions_names[j])
    # random spaces
    value = latent_space_distance(random_embs_1, random_embs_2, fct=fct)
    print('random', value)
    results_df[functions_names[j]]['random'] = value
    # almost identical spaces (just a test)
    value = latent_space_distance(random_embs_1, -random_embs_1, fct=fct)
    print('identical', value)
    results_df[functions_names[j]]['same'] = value

results_df

backbones = ['convnet', 'densenet2']
conds = ['foldlabel_30', 'foldlabel_4', 'no_foldlabel_30', 'no_foldlabel_4']


for backbone in backbones:
    for cond in conds:
        
        print(backbone, cond)

        main_path = f"/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/{backbone}/{cond}"
        save_path = f"/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/#performance_comparisons/nn/{backbone}_{cond}"

        models = [model for model in os.listdir(main_path)
                if os.path.isdir(os.path.join(main_path, model))]
        models.sort()

        print(models)

        functions = [custom_identity, np.sqrt, custom_ln, custom_ln_100]
        functions_names = ['identity', 'sqrt', 'ln', 'ln_100']

        for i in range(len(models)-1):
            for j in range(i+1, len(models)):
                model_path1 = os.path.join(main_path, models[i])
                model_path2 = os.path.join(main_path, models[j])
                
                embs1 = pd.read_csv(model_path1+"/cingulate_ACCpatterns_embeddings/full_embeddings.csv", index_col=0)
                embs2 = pd.read_csv(model_path2+"/cingulate_ACCpatterns_embeddings/full_embeddings.csv", index_col=0)

                values = {}

                dist_mat1 = get_distance_matrix(embs1)
                dist_mat2 = get_distance_matrix(embs2)

                perc_mat1 = get_percentile_matrix(dist_mat1)
                perc_mat2 = get_percentile_matrix(dist_mat2)

                for k,fct in enumerate(functions):
                    values[functions_names[k]] = get_distance(perc_mat1, perc_mat2, fct=fct)
                    print(models[i], models[j], functions_names[k], values[functions_names[k]])
                
                with open(os.path.join(save_path, models[i]+'_'+models[j]+"_nne.json"), 'w') as file:
                    json.dump(values, file)

# sum up the nn distances into a csv

fct = 'sqrt'

saved_path = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/#performance_comparisons/nn"
models = ['densenet2_foldlabel_30', 'densenet2_no_foldlabel_30', 'densenet2_foldlabel_4', 'densenet2_no_foldlabel_4',
'convnet_foldlabel_30', 'convnet_no_foldlabel_30', 'convnet_foldlabel_4', 'convnet_no_foldlabel_4']

mod_conds = [folder for folder in os.listdir(saved_path) if os.path.isdir(os.path.join(saved_path,folder))]

nne_mean = np.zeros((len(mod_conds)//2, 2))

for folder in mod_conds:
    j = 0 if 'convnet' in folder else 1
    if "no" in folder:
        i = 2
    else:
        i = 0
    if '4' in folder:
        i += 1
    
    func = []
    for file in os.listdir(os.path.join(saved_path,folder)):
        if 'nne' in file:

            with open(os.path.join(saved_path, folder, file), 'r') as file_txt:
                file_content = json.load(file_txt)
            func.append(file_content[fct])

    nne_mean[i,j] = np.mean(func)

df = pd.DataFrame(nne_mean, columns=['convnet', 'densenet2'], index=['foldlabel_30', 'foldlabel_4', 'no_foldlabel_30', 'no_foldlabel_4'])
df.to_csv(saved_path+f'/nne_{fct}.csv')
df

### Pointnet nn

main_path = "/neurospin/dico/agaudin/Runs/04_pointnet/Output/pointnet/file_no_30_drop_rate=0.05"
save_path = main_path + "/#nn"

models = [model for model in os.listdir(main_path)
          if ((os.path.isdir(os.path.join(main_path, model)))
              & ("#" not in model))]
models.sort()

print(models)

functions = [custom_ln]
functions_names = ['ln']

for i in range(len(models)-1):
    for j in range(i+1, len(models)):
        model_path1 = os.path.join(main_path, models[i])
        model_path2 = os.path.join(main_path, models[j])
        
        embs1 = pd.read_csv(model_path1+"/cingulate_ACCpatterns_embeddings/full_embeddings.csv", index_col=0)
        embs2 = pd.read_csv(model_path2+"/cingulate_ACCpatterns_embeddings/full_embeddings.csv", index_col=0)

        values = {}

        dist_mat1 = get_distance_matrix(embs1)
        dist_mat2 = get_distance_matrix(embs2)

        perc_mat1 = get_percentile_matrix(dist_mat1)
        perc_mat2 = get_percentile_matrix(dist_mat2)

        for k,fct in enumerate(functions):
            values[functions_names[k]] = get_distance(perc_mat1, perc_mat2, fct=fct)
            print(models[i], models[j], functions_names[k], values[functions_names[k]])
        
        with open(os.path.join(save_path, models[i]+'_'+models[j]+"_nne.json"), 'w') as file:
            json.dump(values, file)

main_path = "/neurospin/dico/agaudin/Runs/04_pointnet/Output/pointnet"
folders = [folder for folder in os.listdir(main_path)
           if 'file' in folder]

for folder in folders:
    path = os.path.join(main_path, folder) + '/#nn'
    mean_nn = []
    for filename in os.listdir(path):
        if 'nne' in filename:
            with open(path+'/'+filename, 'r') as file:
                nne = json.load(file)
                mean_nn.append(nne['ln'])
    mean_nn = np.mean(mean_nn)
    with open(path+'/mean_nn.txt', 'w') as save_file:
        save_file.write(str(mean_nn))

## Compare the results 

saved_path = "/neurospin/dico/agaudin/Runs/03_monkeys/Output/analysis_folders/#performance_comparisons/nn"

id_df = pd.read_csv(saved_path+'/nne_identity.csv', index_col=0)
sqrt_df = pd.read_csv(saved_path+'/nne_sqrt.csv', index_col=0)
ln_df = pd.read_csv(saved_path+'/nne_ln.csv', index_col=0)
ln_100_df = pd.read_csv(saved_path+'/nne_ln_100.csv', index_col=0)


nn_df = pd.read_csv(saved_path+'/mean_median_nn.csv', index_col=0)

print("id")
print(id_df)
print()

print("sqrt")
print(sqrt_df)
print()

print("ln")
print(ln_df)
print()

print("ln_100")
print(ln_100_df)
print()

print("nn")
print(nn_df)

dfs = [id_df, sqrt_df, ln_df, ln_100_df, nn_df]

cond = 'no_foldlabel_4'

for df in dfs:
    ratio = df['convnet'][cond]/df['densenet2'][cond]
    print(ratio)

dfs = [id_df, sqrt_df, ln_df, ln_100_df, nn_df]

model = 'convnet'
cond1 = 'no_foldlabel_30'
cond2 = 'foldlabel_30'

for df in dfs:
    ratio = df[model][cond1]/df[model][cond2]
    print(ratio)

fct has almost no impact on the predictions: the absolute values might be different, but the ratio remains the same.
